test(range): migrate events test
